Fixed how exports work
